AeroGarden Front End Dev Challenge
Hey Chris, here is the result of my code challenge parts 1 through 3. I must admit that I had to do some research for some of the tasks, but it was fun to pull it all together. For part 3, I chose to make a simple .html file vs. a .js file, just so you can test/see the result right away (I didn't think a script src link to a challenge.js file was needed for this simple demo hehe). I have also included a screen shot that shows the render/output of my code in part 1. Thanks again, and feedback is welcome! -Brian-